[KSP-115] refactor: 파일 구조 수정 및 components 폴더로 이동
